chore: migrate commitlint config to TS
